+++ MonkeyDelivery/Src/Enemies/Cat.cpp
@@ -38,8 +38,8 @@ void Cat::draw()
 {
 	if (isAlive()) {
 		auto pos = getCollider();
-		pos.x -= game->getCamera()->getCameraPosition().getX();
-		pos.y -= game->getCamera()->getCameraPosition().getY();
+		pos.x -= (int)game->getCamera()->getCameraPosition().getX();
+		pos.y -= (int)game->getCamera()->getCameraPosition().getY();
 		if (collided && (game->getPlayer()->isUsingFlashLight() || game->getPlayer()->isUsingLantern())) 
 			setTexture(catSS_Death);
