+++ MonkeyDelivery/Src/Enemies/Plant.cpp
@@ -56,8 +56,8 @@ void Plant::draw()
 {
 	if (isAlive()) {
 		auto pos = getCollider();
-		pos.x -= game->getCamera()->getCameraPosition().getX();
-		pos.y -= game->getCamera()->getCameraPosition().getY();
+		pos.x -= (int)game->getCamera()->getCameraPosition().getX();
+		pos.y -= (int)game->getCamera()->getCameraPosition().getY();
 		if (collided && (game->getPlayer()->isUsingFlashLight() || game->getPlayer()->isUsingLantern())) 
 			setTexture(plantSS_Death);
