+++ MonkeyDelivery/Src/Items/Rock.h
@@ -2,8 +2,8 @@
 #include "../Logic/ColliderTile.h"
 
 class Rock : public ColliderTile {
-	int w=75.0;
-	int h=75.0;
+	double w=75.0;
+	double h=75.0;
 	int levelToDestroy_;
 public:
 	Rock(Game* game, int x, int y,int level);
